+++ sensor/Dust25S.cpp
@@ -3,8 +3,8 @@
 #include <math.h>
 namespace Sensor{
 
-float Dust25S::trasmute(float rawMeasure) const  {
-    float concentrationPM25 = a * rawMeasure + b;
+double Dust25S::trasmute(double rawMeasure) const  {
+    double concentrationPM25 = a * rawMeasure + b;
     return concentrationPM25;
 };
 Dust25S::~Dust25S(){};
@@ -14,11 +14,11 @@ Dust25S::Dust25S(   const std::string name,
                 const  std::string longDesc,
                 const  std::string xAxisLabel,
                 const  unsigned int simulationSpan,
-                const  float sensibility,
-                const  float maxMeasurable,
-                const  float minMeasurable,
-                const float a,
-                const float b) 
+                const  double sensibility,
+                const  double maxMeasurable,
+                const  double minMeasurable,
+                const double a,
+                const double b) 
                 : AbstractSensor(name, "Resistore a impasto di carbone", shortDesc, longDesc, xAxisLabel, "PM 2.5 [µg/m³]" ,simulationSpan, sensibility, maxMeasurable, minMeasurable ),
                 a(a), b(b){};
 
@@ -27,15 +27,15 @@ void Dust25S::accept(VisitorInterface& visitor){
 };
 
 // getter
-float Dust25S::getA() const {return a;};
-float Dust25S::getB() const {return b;};
+double Dust25S::getA() const {return a;};
+double Dust25S::getB() const {return b;};
 
 // setter
-void Dust25S::setA( float A){ 
+void Dust25S::setA( double A){ 
     a=A ;
 notifyAllObservers(this);
 }; 
-void Dust25S::setB( float B){ 
+void Dust25S::setB( double B){ 
     b=B ;
 notifyAllObservers(this);
 };
